Move the format helpers to l.lib.format.
