+++ landscape/manager/plugin.py
@@ -1,7 +1,7 @@
 from twisted.internet.defer import maybeDeferred
 
+from landscape.lib.format import format_object
 from landscape.lib.log import log_failure
-from landscape.log import format_object
 from landscape.broker.client import BrokerClientPlugin
 
 # Protocol messages! Same constants are defined in the server.
